Added registration screen
Added the registration screen that can now be accessed when pressing the "register button on the welcome screen. Creates a new user by entering a name, id, password, and account type and clicking the register button. Verifies if a user with the login id is  already created. The cancel button returns to the welcome screen without saving the credentials. The login screen no longer has hard coded information and now checks the user list to authenticate if the credentials are valid.